refactor: move post orders on new encounter in common lib
