+++ packages/esm-patient-common-lib/src/orders/postOrders.ts
@@ -1,8 +1,58 @@
-import { openmrsFetch, restBaseUrl } from '@openmrs/esm-framework';
+import { openmrsFetch, type OpenmrsResource, parseDate, restBaseUrl, type Visit } from '@openmrs/esm-framework';
 import { getPatientUuidFromUrl } from '../get-patient-uuid-from-url';
 import { type OrderBasketStore, orderBasketStore } from './store';
-import { type ExtractedOrderErrorObject, type OrderBasketItem, type OrderPost } from './types';
-import { type OrderErrorObject } from './types';
+import { type ExtractedOrderErrorObject, type OrderBasketItem, type OrderErrorObject, type OrderPost } from './types';
+
+export async function postOrdersOnNewEncounter(
+  patientUuid: string,
+  orderEncounterType: string,
+  activeVisit: Visit | null,
+  sessionLocationUuid: string,
+  abortController?: AbortController,
+) {
+  const now = new Date();
+  const visitStartDate = parseDate(activeVisit?.startDatetime);
+  const visitEndDate = parseDate(activeVisit?.stopDatetime);
+  let encounterDate: Date;
+  if (!activeVisit || (visitStartDate < now && (!visitEndDate || visitEndDate > now))) {
+    encounterDate = now;
+  } else {
+    console.warn(
+      'create Encounter received an active visit that is not currently active. This is a programming error. Attempting to place the order using the visit start date.',
+    );
+    encounterDate = visitStartDate;
+  }
+
+  const { items, postDataPrepFunctions }: OrderBasketStore = orderBasketStore.getState();
+  const patientItems = items[patientUuid];
+
+  const orders: Array<OrderPost> = [];
+
+  Object.entries(patientItems).forEach(([grouping, groupOrders]) => {
+    groupOrders.forEach((order) => {
+      orders.push(postDataPrepFunctions[grouping](order, patientUuid, null));
+    });
+  });
+
+  const encounterPostData = {
+    patient: patientUuid,
+    location: sessionLocationUuid,
+    encounterType: orderEncounterType,
+    encounterDatetime: encounterDate,
+    visit: activeVisit?.uuid,
+    obs: [],
+    orders,
+  };
+
+  return openmrsFetch<OpenmrsResource>(`${restBaseUrl}/encounter`, {
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    method: 'POST',
+    body: encounterPostData,
+    signal: abortController?.signal,
+  }).then((res) => res?.data?.uuid);
+}
 
 export async function postOrders(encounterUuid: string, abortController: AbortController) {
   const patientUuid = getPatientUuidFromUrl();
