+++ packages/esm-patient-orders-app/src/order-basket/order-basket.workspace.tsx
@@ -4,14 +4,15 @@ import { type TFunction, useTranslation } from 'react-i18next';
 import { ActionableNotification, Button, ButtonSet, InlineNotification } from '@carbon/react';
 import { ExtensionSlot, showModal, showSnackbar, useConfig, useLayoutType, useSession } from '@openmrs/esm-framework';
 import {
-  type OrderBasketItem,
   type DefaultPatientWorkspaceProps,
+  type OrderBasketItem,
   postOrders,
+  postOrdersOnNewEncounter,
   useOrderBasket,
   useVisitOrOfflineVisit,
 } from '@openmrs/esm-patient-common-lib';
 import { type ConfigObject } from '../config-schema';
-import { saveOrdersWithNewEncounter, useMutatePatientOrders, useOrderEncounter } from '../api/api';
+import { useMutatePatientOrders, useOrderEncounter } from '../api/api';
 import styles from './order-basket.scss';
 
 const OrderBasket: React.FC<DefaultPatientWorkspaceProps> = ({
@@ -57,7 +58,7 @@ const OrderBasket: React.FC<DefaultPatientWorkspaceProps> = ({
     // If there's no encounter present, create an encounter along with the orders.
     if (!orderEncounterUuid) {
       try {
-        await saveOrdersWithNewEncounter(
+        await postOrdersOnNewEncounter(
           patientUuid,
           config?.orderEncounterType,
           activeVisitRequired ? activeVisit : null,
@@ -70,11 +71,11 @@ const OrderBasket: React.FC<DefaultPatientWorkspaceProps> = ({
         closeWorkspaceWithSavedChanges();
         showOrderSuccessToast(t, orders);
       } catch (e) {
+        console.error(e);
         setCreatingEncounterError(
           e.responseBody.error.message ||
             t('tryReopeningTheWorkspaceAgain', 'Please try launching the workspace again'),
         );
-        console.error(e);
       }
     } else {
       const erroredItems = await postOrders(orderEncounterUuid, abortController);
